no-op systemd for non-linux systems
