+++ pkg/health/systemd_test.go
@@ -1,3 +1,5 @@
+//go:build linux
+
 package health
 
 import (
